backend/requirements.txt
flask==2.3.3
flask-login==0.6.3
flask-jwt-extended==4.5.3
flask-sqlalchemy==3.0.5
flask-migrate==4.0.5
flask-wtf==1.1.1
flask-cors==4.0.0
flask-dance==7.0.0
python-dotenv==1.0.0
sqlalchemy==2.0.21
psycopg2-binary==2.9.7
werkzeug==2.3.7
wtforms==3.0.1
email-validator==2.0.0
bleach==6.0.0
